Z/H/CHGenerator: Remove gen timestamp for reproducibility
The change remove the timestamp to enable reproducible builds.

It will help to automate xml updates in consuming projects, they can
be re-generated at buildtime, this timestamp info is adding noise.

By the way the ref of the tool is added.

One improvement to consider is to add git-describe of inputs files (and tool) in header, but it can also generate noise too, we lived without that so far.

Origin: https://github.com/Z-Wave-Alliance/z-wave-tools-core/pull/44
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack/issues/88
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-xml-tools/issues/4
Relate-to: https://github.com/Z-Wave-Alliance/zwave_xml/issues/103
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>